clips are prozentual
git-svn-id: https://blinkendroid.googlecode.com/svn/trunk@185 ded441ff-28c8-dbd9-35dd-4b5d9abe79e0

+++ blinkendroid/src/org/cbase/blinkendroid/network/BlinkendroidListener.java
@@ -22,7 +22,7 @@ package org.cbase.blinkendroid.network;
  */
 public interface BlinkendroidListener {
 
-    void clip(int startX, int startY, int endX, int endY);
+    void clip(float startX, float startY, float endX, float endY);
 
     void serverTime(long serverTime);
